Initial import
In this commit, I created a blank Visual Studio 2019 solution. I also wrote the OpenApi document (datetime.yaml) and I added the openapi-generator-cli.
For an OpenApi-first development methodology, it is necessary to have the OpenApi document written first. I will use it to generate a server and a client stub in the future.