Modify server generation to have proper generated code
In this commit I found out that I can provide parameters to the aspnetcore generator.
To read more about these parameters, take a look at this link:
https://github.com/OpenAPITools/openapi-generator/blob/master/docs/generators/aspnetcore.md

The aspnetcore generator targets a preview version of ASP.NET Core 3.0, but I don't mind it, since I am going to copy the generated code to a custom project.